+++ src/Database/Element/Behavior/IndexBehavior.php
@@ -3,7 +3,9 @@
 namespace Phoenix\Database\Element\Behavior;
 
 use Phoenix\Database\Element\Index;
+use Phoenix\Database\Element\IndexColumn;
 use Phoenix\Database\Element\MigrationTable;
+use Phoenix\Exception\InvalidArgumentValueException;
 
 trait IndexBehavior
 {
@@ -12,17 +14,16 @@ trait IndexBehavior
     private $indexesToDrop = [];
 
     /**
-     * @param string|array $columns name(s) of column(s)
+     * @param string|array|IndexColumn|IndexColumn[] $columns name(s) of column(s) or IndexColumn instance(s)
      * @param string $type type of index (unique, fulltext) default ''
      * @param string $method method of index (btree, hash) default ''
      * @param string $name name of index
      * @return MigrationTable
+     * @throws InvalidArgumentValueException if index type or index method is not allowed
      */
     public function addIndex($columns, string $type = Index::TYPE_NORMAL, string $method = Index::METHOD_DEFAULT, string $name = ''): MigrationTable
     {
-        if (!is_array($columns)) {
-            $columns = [$columns];
-        }
+        $columns = $this->createIndexColumns($columns);
         $index = new Index($columns, $this->createIndexName($columns, $name), $type, $method);
         $this->indexes[] = $index;
         return $this;
@@ -38,12 +39,11 @@ trait IndexBehavior
 
     /**
      * @param string|array $columns
+     * @return MigrationTable
      */
     public function dropIndex($columns): MigrationTable
     {
-        if (!is_array($columns)) {
-            $columns = [$columns];
-        }
+        $columns = $this->createIndexColumns($columns);
         $indexName = $this->createIndexName($columns);
         return $this->dropIndexByName($indexName);
     }
@@ -59,13 +59,48 @@ trait IndexBehavior
         return $this->indexesToDrop;
     }
 
+    /**
+     * @param IndexColumn[] $columns
+     * @param string $name
+     * @return string
+     */
     private function createIndexName(array $columns, string $name = ''): string
     {
         if ($name) {
             return $name;
         }
 
-        return 'idx_' . $this->getName() . '_' . implode('_', $columns);
+        $nameParts = [
+            'idx',
+            $this->getName(),
+        ];
+        foreach ($columns as $column) {
+            $nameParts[] = $column->getName();
+            $columnSettings = $column->getSettings()->getNonDefaultSettings();
+            ksort($columnSettings);
+            foreach ($columnSettings as $setting => $value) {
+                $nameParts[] = substr($setting, 0, 1) . $value;
+            }
+        }
+        return strtolower(implode('_', $nameParts));
+    }
+
+    /**
+     * @param array|string $columns
+     * @return IndexColumn[]
+     */
+    private function createIndexColumns($columns): array
+    {
+        if (!is_array($columns)) {
+            $columns = [$columns];
+        }
+
+        $columnList = [];
+        foreach ($columns as $column) {
+            $column = $column instanceof IndexColumn ? $column : new IndexColumn($column);
+            $columnList[] = $column;
+        }
+        return $columnList;
     }
 
     abstract public function getName();