+++ testing_migrations/phoenix/20150908065215_add_data.php
@@ -69,12 +69,15 @@ class AddData extends AbstractMigration
 
         $this->insert('table_5', [
             [
+                'id' => 1,
                 'title' => 'Item 1',
             ],
             [
+                'id' => 2,
                 'title' => 'Item 2',
             ],
             [
+                'id' => 3,
                 'title' => 'Item 3',
             ],
         ]);